Added logic for index in the NativeElementFinderService
When a ComponentList is created, it assigns ElementIndex to the components, but because of the logic in GetWebElement() inside Component class, it will try to 'find' it again with incorrect index.
The logic for 're-finding' elements should be removed altogether from the playwright module, because ILocator does that behind the scenes. ILocator only holds the logic for finding an element, not a reference to the element itself.
That's the reason why in Playwright there is no such thing as 'StaleElementException', compared to Selenium.