Lighthouse Reports Workflow Plugins update
One can run a lighthouse analysis only with LighthouseAnalysisRunAttribute, so there is need to validate if in the config file the lighthouse is enabled or not.
Changes to these files are made due to the fact that if one forgets to set 'isEnabled' in LighthouseSettings to 'false', even if the tests are not lighthouse analysis, this plugin will attach the latest reports (if any) to them.